+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog references the relevant changes (new features, changes and bugs) 
   * Fixed accessor typo in Enabled entity attribute.
   * Added Published entity attribute.
   * Restructured navigation menu.
+  * Added Published entity attribute to Service entity.
 
 
 # CHANGELOG for 0.3.0